refactor: 프로젝트 패키징
