+++ src/main/java/lotto/util/RandomGenerator.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.util;
+
+import lotto.domain.Lotto;
 
 import java.util.Collections;
 import java.util.List;
