+++ src/test/java/lotto/LottoResultsTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.LottoRank;
+import lotto.domain.LottoResults;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
