+++ src/test/java/lotto/LottoServiceTest.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.domain.Lottos;
+import lotto.domain.Money;
+import lotto.service.LottoService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
