+++ src/test/java/lotto/MoneyTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.Money;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
